Refactor index.jsp and details.jsp

+++ src/main/java/com/codeup/adlister/controllers/AdDetailsServlet.java
@@ -10,6 +10,10 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
+import java.io.PrintWriter;
+
+import static java.lang.System.in;
+import static java.lang.System.out;
 
 @WebServlet("/ads/details")
 public class AdDetailsServlet extends HttpServlet {
@@ -17,16 +21,25 @@ public class AdDetailsServlet extends HttpServlet {
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
         long id = Long.parseLong(req.getParameter("id"));
         Ad ad = DaoFactory.getAdsDao().findById(id);
-        User user = DaoFactory.getUsersDao().findById(ad.getUserId());
         req.getSession().setAttribute("ad", ad);
-        req.getSession().setAttribute("user", user);
         req.getRequestDispatcher("/WEB-INF/details.jsp").forward(req, resp);
     }
 
     @Override
     protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
         Ad ad = (Ad) req.getSession().getAttribute("ad");
-        DaoFactory.getAdsDao().delete(ad);
-        resp.sendRedirect("/ads");
+        User user = (User) req.getSession().getAttribute("user");
+        boolean incorrectUser;
+        if (user == null) {
+            resp.sendRedirect("/login");
+        } else if (ad.getUserId() == user.getId()) {
+            DaoFactory.getAdsDao().delete(ad);
+            resp.sendRedirect("/ads");
+        } else {
+            resp.setContentType("text/html");
+            PrintWriter out = resp.getWriter();
+            resp.sendRedirect("/ads");
+            out.println("<script>alert('Unable to delete ads from other users.')</script>");
+        }
     }
 }
